Factor muxer for improved extensibility.
Updated version of https://github.com/PolymerLabs/arcs/pull/1018.

Changes for social that make use of this new structure to follow
separately.
