+++ runtime/loader.js
@@ -14,6 +14,7 @@ import fetch from './fetch-web.js';
 import assert from '../platform/assert-web.js';
 import particle from './particle.js';
 import DomParticle from './dom-particle.js';
+import MultiplexerDomParticle from './multiplexer-dom-particle.js';
 import TransformationDomParticle from './transformation-dom-particle.js';
 import JsonldToManifest from './converters/jsonldToManifest.js';
 const html = (strings, ...values) => (strings[0] + values.map((v, i) => v + strings[i + 1]).join('')).trim();
@@ -89,7 +90,7 @@ class Loader {
   }
 
   unwrapParticle(particleWrapper) {
-    return particleWrapper({particle, Particle: particle.Particle, DomParticle, TransformationDomParticle, html});
+    return particleWrapper({particle, Particle: particle.Particle, DomParticle, TransformationDomParticle, MultiplexerDomParticle, html});
   }
 
 }
